refactor: move client focus functions from objects/client.c to focus.c
Extract 14 focus-related engine functions from objects/client.c into
focus.c: client_unfocus_internal, client_unfocus, client_ban_unfocus,
client_ban, client_unban, client_ignore/restore_enterleave_events,
client_focus_update, client_focus, client_focus_refresh,
client_border_refresh, client_geometry_refresh, client_refresh,
client_destroy_later.

Promoted client_set_hidden from static to public (needed by
client_ban in focus.c). objects/client.c is now purely Lua bindings
for focus operations.

objects/client.c: 5399 -> 5064 lines (-335)
focus.c: 240 -> 577 lines (+337)
